vendor/containers/image: add workaround for Quay
Unfortunately Quay is a bit broken with skopeo at the moment because it
returns a 500 instead of a 400 when it receives an unknown/invalid
manifest. This causes skopeo to bail out rather than trying a different
manifest type after first trying to push the preferred OCI manifest
type. I've reported this upstream with Quay, but until it's fixed there,
let's just avoid pushing OCI manifests entirely, as we don't strictly
speaking have a need for them.